modificaçoes no request.js e TripDetailsForm.js
